deployment_requirements.txt
django==5.2
djangorestframework==3.16.0
dj-database-url==2.3.0
mysqlclient==2.2.4
pyMySQL==1.1.0
numpy==1.26.3
pandas==2.2.1
scikit-learn==1.4.1
joblib==1.3.2
matplotlib==3.8.2
statsmodels==0.14.1
tensorflow==2.15.0
requests==2.31.0
trafilatura==1.6.3
geopy==2.4.1
shapely==2.0.3
beautifulsoup4==4.12.3
python-dateutil==2.8.2
pytz==2024.1
gunicorn==21.2.0
whitenoise==6.6.0